+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update design pattern
 - Modify the setting note config `doNotShowJournalPreview` to inlcude even GM not only the players
 - Integration with
+- Set default value of the module setting `noteGM` to true
 
 ## [1.5.16]